build with cmake
