--- BaseMessage.h
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace MessageBus {
-
-struct BaseMessage {
-    virtual ~BaseMessage() = default;
-};
-
-}
