+++ include/MessageBus/BaseMessage.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace MessageBus {
+
+struct BaseMessage {
+    virtual ~BaseMessage() = default;
+};
+
+}
